Features added and updated
The ability to load to and from JSON files has been added. The framework
to use 'share ids' to share swot analysis with other people easily has
been added, but the feature is not complete. Fixed urls.py so that the
website has a way to access access_app.html